requirements.txt
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
pandas==1.1.0
Pillow==7.2.0
streamlit==0.64.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0